Bug: `GET` Redirects from "modal" to "basic"
According to the path configuration file, the client handles `GET
/navigations` requests with the "basic" stack, and `GET /modal/new`
requests with the "modal" stack.

This commit modifies the `modals#new` action to redirect to
`navigations#show` when the `?redirect` query parameter is present.

This `GET` to `GET` redirect causes issues on the client, since at
request-time, the client **thinks** its handling a "modal" navigation,
but at response-time, the client is **served** a route that it'd handle
as a "basic"` navigation.

The resulting behavior is a bug through the following sequence:

1. present the "modal" stack by animating up from the bottom of the
   screen
2. render an empty screen
3. dismiss the "modal" stack
4. pop the entirety of the "basic" stack
5. replace the root with the `navigations#show` response

Once the navigation completes, the old root screen is entirely
unavailable. There is no navigation "Back" button, gestures don't pop
off the stack, and pull to refresh re-requests the `navigations#show`
page.